requirements.txt
streamlit
requests
pandas
joblib
mysql-connector-python
scikit-learn
tensorflow
keras


